mensajes, menu y usuario en rol, modulo, usuario, nivel puesto y puesto

+++ src/com/shuldevelop/controller/ModuloController.java
@@ -5,6 +5,9 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -12,10 +15,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.shuldevelop.model.Modulo;
+import com.shuldevelop.model.Usuario;
 import com.shuldevelop.model.validator.ModuloValidator;
 import com.shuldevelop.service.ModuloService;
+import com.shuldevelop.service.UsuarioService;
 
 @Controller
 public class ModuloController {
@@ -23,10 +29,27 @@ public class ModuloController {
 	@Autowired
 	private ModuloService moduloService;
 	
+	@Autowired
+	private UsuarioService usuarioService;
+	
 	private ModuloValidator moduloValidator;
 	
 	public ModuloController() {
 		this.moduloValidator = new ModuloValidator();
+	}
+	
+	public Usuario getUsuario() {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		
+		UserDetails userDetails = (UserDetails) auth.getPrincipal();
+		
+		return usuarioService.findUserByUsername(userDetails.getUsername());
+	}
+	
+	public List<Modulo> getModulos() {
+		
+		return moduloService.getAllModuloByRol(getUsuario().getRol().getId());
+		
 	}
 
 	@RequestMapping(value = "/modulo/index", method = RequestMethod.GET)
@@ -38,6 +61,8 @@ public class ModuloController {
 		
 		mav.setViewName("modulo/index");
 		mav.addObject("moduloList", listModulo);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		return mav;
 	}
@@ -52,6 +77,8 @@ public class ModuloController {
 		mav.setViewName("modulo/add");
 		mav.addObject("Modulo", new Modulo());
 		mav.addObject("moduloList", listModulo);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		return mav;
 	}
@@ -60,7 +87,8 @@ public class ModuloController {
 	public ModelAndView addModulo(
 			@ModelAttribute("Modulo") Modulo u,
 			BindingResult result,
-			SessionStatus status
+			SessionStatus status,
+			final RedirectAttributes redirectAttributes
  			) {
 		
 		this.moduloValidator.validate(u, result);
@@ -73,6 +101,8 @@ public class ModuloController {
 			mav.setViewName("modulo/add");
 			mav.addObject("Modulo", u);
 			mav.addObject("moduloList", listModulo);
+			mav.addObject("Usuario", getUsuario());
+			mav.addObject("modulos", getModulos());
 			
 			return mav;
 		}
@@ -80,6 +110,8 @@ public class ModuloController {
 		u.setDependencia( moduloService.getModulo( u.getDependencia().getId() ) );
 		
 		moduloService.add(u);
+		
+		redirectAttributes.addFlashAttribute("messageSuccess", "El modulo ha sido agregado exitosamente.");
 		
 		return new ModelAndView("redirect:/modulo/index.html");
 	}
@@ -97,6 +129,8 @@ public class ModuloController {
 		mav.setViewName("modulo/edit");
 		mav.addObject("Modulo", modulo);
 		mav.addObject("moduloList", listModulo);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		return mav;
 	}
@@ -106,7 +140,8 @@ public class ModuloController {
 			@ModelAttribute("Modulo") Modulo u,
 			BindingResult result,
 			SessionStatus status,
-			HttpServletRequest request
+			HttpServletRequest request,
+			final RedirectAttributes redirectAttributes
  			) {
 		
 		this.moduloValidator.validate(u, result);
@@ -119,6 +154,8 @@ public class ModuloController {
 			mav.setViewName("modulo/edit");
 			mav.addObject("Modulo", u);
 			mav.addObject("moduloList", listModulo);
+			mav.addObject("Usuario", getUsuario());
+			mav.addObject("modulos", getModulos());
 			
 			return mav;
 		}
@@ -127,15 +164,30 @@ public class ModuloController {
 		
 		moduloService.edit(u);
 		
+		redirectAttributes.addFlashAttribute("messageSuccess", "El modulo ha sido editado exitosamente.");
+		
 		return new ModelAndView("redirect:/modulo/index.html");
 	}
 	
 	@RequestMapping(value = "/modulo/delete", method = RequestMethod.GET)
-	public ModelAndView deletePuesto(HttpServletRequest request) {
+	public ModelAndView deletePuesto(
+			HttpServletRequest request,
+			final RedirectAttributes redirectAttributes) {
 		
 		int id = Integer.parseInt(request.getParameter("id"));
 		
-		moduloService.delete(id);
+		Modulo modulo = moduloService.getModulo(id);
+		
+		if ( modulo.getModulos().size() > 0 || moduloService.hasModuloRolPermiso( modulo.getId() )) {
+			
+			redirectAttributes.addFlashAttribute("messageError", "El modulo posee dependencia no se puede eliminar.");
+			
+		} else {
+			
+			moduloService.delete(id);
+			redirectAttributes.addFlashAttribute("messageSuccess", "El modulo ha sido eliminado exitosamente.");
+			
+		}
 		
 		return new ModelAndView("redirect:/modulo/index.html");
 	}
