+++ src/com/shuldevelop/controller/PuestoController.java
@@ -6,6 +6,9 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -13,12 +16,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.shuldevelop.model.Modulo;
 import com.shuldevelop.model.NivelPuesto;
 import com.shuldevelop.model.Puesto;
+import com.shuldevelop.model.Usuario;
 import com.shuldevelop.model.validator.PuestoValidator;
+import com.shuldevelop.service.ModuloService;
 import com.shuldevelop.service.NivelPuestoService;
 import com.shuldevelop.service.PuestoService;
+import com.shuldevelop.service.UsuarioService;
 
 @Controller
 public class PuestoController {
@@ -29,10 +37,30 @@ public class PuestoController {
 	@Autowired
 	private NivelPuestoService nivelPuestoService;
 	
+	@Autowired
+	private UsuarioService usuarioService;
+	
+	@Autowired
+	private ModuloService moduloService;
+	
 	private PuestoValidator puestoValidator;
 
 	public PuestoController() {
 		puestoValidator = new PuestoValidator();
+	}
+	
+	public Usuario getUsuario() {
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		
+		UserDetails userDetails = (UserDetails) auth.getPrincipal();
+		
+		return usuarioService.findUserByUsername(userDetails.getUsername());
+	}
+	
+	public List<Modulo> getModulos() {
+		
+		return moduloService.getAllModuloByRol(getUsuario().getRol().getId());
+		
 	}
 	
 	@RequestMapping(value = "/puesto/index", method = RequestMethod.GET)
@@ -44,6 +72,8 @@ public class PuestoController {
 		
 		mav.setViewName("puesto/index");
 		mav.addObject("puestoList", listPuesto);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		return mav;
 	}
@@ -58,6 +88,8 @@ public class PuestoController {
 		mav.setViewName("puesto/add");
 		mav.addObject("Puesto", new Puesto());
 		mav.addObject("tipoPuestoList", listTipoPuesto);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		
 		return mav;
@@ -67,7 +99,8 @@ public class PuestoController {
 	public ModelAndView addPuesto(
 			@ModelAttribute("Puesto") Puesto u,
 			BindingResult result,
-			SessionStatus status
+			SessionStatus status,
+			final RedirectAttributes redirectAttributes
  			) {
 		
 		this.puestoValidator.validate(u, result);
@@ -80,6 +113,8 @@ public class PuestoController {
 			mav.setViewName("puesto/add");
 			mav.addObject("Puesto", u);
 			mav.addObject("tipoPuestoList", listTipoPuesto);
+			mav.addObject("Usuario", getUsuario());
+			mav.addObject("modulos", getModulos());
 			
 			
 			return mav;
@@ -93,6 +128,8 @@ public class PuestoController {
 				u.getNivelPuesto().getId()).getNumeroNivel());
 		
 		puestoService.add(u);
+		
+		redirectAttributes.addFlashAttribute("messageSuccess", "El puesto ha sido agregado exitosamente.");
 		
 		return new ModelAndView("redirect:/puesto/index.html");
 	}
@@ -110,6 +147,8 @@ public class PuestoController {
 		mav.setViewName("puesto/edit");
 		mav.addObject("Puesto", puesto);
 		mav.addObject("tipoPuestoList", listTipoPuesto);
+		mav.addObject("Usuario", getUsuario());
+		mav.addObject("modulos", getModulos());
 		
 		return mav;
 	}
@@ -119,7 +158,8 @@ public class PuestoController {
 			@ModelAttribute("Puesto") Puesto u,
 			BindingResult result,
 			SessionStatus status,
-			HttpServletRequest request
+			HttpServletRequest request,
+			final RedirectAttributes redirectAttributes
  			) {
 		
 		this.puestoValidator.validate(u, result);
@@ -132,9 +172,13 @@ public class PuestoController {
 			mav.setViewName("nivel_puesto/edit");
 			mav.addObject("Puesto", u);
 			mav.addObject("tipoPuestoList", listTipoPuesto);
+			mav.addObject("Usuario", getUsuario());
+			mav.addObject("modulos", getModulos());
 			
 			return mav;
 		}
+		
+		redirectAttributes.addFlashAttribute("messageSuccess", "El puesto ha sido editado exitosamente.");
 		
 		puestoService.edit(u);
 		
@@ -142,11 +186,23 @@ public class PuestoController {
 	}
 	
 	@RequestMapping(value = "/puesto/delete", method = RequestMethod.GET)
-	public ModelAndView deletePuesto(HttpServletRequest request) {
+	public ModelAndView deletePuesto(
+			HttpServletRequest request,
+			final RedirectAttributes redirectAttributes
+			) {
 		
 		int id = Integer.parseInt(request.getParameter("id"));
 		
-		puestoService.delete(id);
+		if ( puestoService.hasPuestoEmpleado(id) ) {
+			
+			redirectAttributes.addFlashAttribute("messageError", "El puesto posee dependencia no se puede eliminar.");
+			
+		} else {
+		
+			puestoService.delete(id);
+			redirectAttributes.addFlashAttribute("messageSuccess", "El puesto ha sido eliminado exitosamente.");
+			
+		}
 		
 		return new ModelAndView("redirect:/puesto/index.html");
 	}
